simple jobified delaunay triangulation
