--- Runtime/Delaunay.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ef895fd19d904c448ae0a230164704de
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
