--- Runtime/Delaunay/DelaunayTriangulator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 80cfcd51b1bc1bd41b8e07996c57bea5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
