--- Runtime/Delaunay/Triangle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 90d7d5a409541ba4b916978750702dcc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
