+++ Runtime/DelaunayTriangulation/DelaunayTriangulation.Primitive.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2225fd6d6dff70745bb477cb267c6551
+guid: 2b875d3c35744824b95e6e96a9ec9f14
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
