--- Runtime/Triangulation/Triangulator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7b37f4b855b80574eae4d2b7523d5f5d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
